Introduce dynamic text styles and various font modifiers (#186)
* Implement dynamic text styles (and various font-related modifiers)

* Fix AppKitBackend tests (broke due to changed default font size)

* Update Font.monospaced(_:) modifier with default parameter value

* Fix AppKitBackend post-rebase and revert CounterApp

* Fix UIKitBackend compilation error (post TextEditor rebase)

+++ Sources/Gtk3/Utility/CSS/CSSProperty.swift
@@ -58,7 +58,7 @@ public struct CSSProperty: Equatable {
         CSSProperty(key: "min-height", value: "\(height)px")
     }
 
-    public static func fontSize(_ size: Int) -> CSSProperty {
+    public static func fontSize(_ size: Double) -> CSSProperty {
         CSSProperty(key: "font-size", value: "\(size)px")
     }
